requirements.txt
pip==19.3.1
config-client==0.4.0
future==0.17.1
Flask==1.1.1
grpcio
grpcio-tools
tensorflow==1.15.0
tensorflow-estimator==1.15.1
tensorflow-hub==0.6.0
tensorflow-metadata==0.14.0
tensorflow-serving-api==1.15.0
pandas==0.25.3
WTForms==2.2.1
